+++ pywatts/core/base_step.py
@@ -121,19 +121,21 @@ class BaseStep(ABC):
     def _pack_data(self, start, end, buffer_element=None, return_all=False):
         # Provide requested data
         time_index = _get_time_indexes(self.buffer)
-        if end and start and end > start:
+        if start:
             index = list(self.buffer.values())[0].indexes[time_index[0]]
-            start = max(index[0], start.to_numpy())
+            start = start.to_numpy()
+            # If end is not set, all values should be considered. Thus we add a small timedelta to the last index entry.
+            end = end.to_numpy() if end is not None else (index[-1] + pd.Timedelta(nanoseconds=1)).to_numpy()
             # After sel copy is not needed, since it returns a new array.
             if buffer_element is not None:
                 return self.buffer[buffer_element].sel(
-                    **{time_index[0]: index[(index >= start) & (index < end.to_numpy())]})
+                    **{time_index[0]: index[(index >= start) & (index < end)]})
             elif return_all:
-                return {key: b.sel(**{time_index[0]: index[(index >= start) & (index < end.to_numpy())]}) for
+                return {key: b.sel(**{time_index[0]: index[(index >= start) & (index < end)]}) for
                         key, b in self.buffer.items()}
             else:
                 return list(self.buffer.values())[0].sel(
-                    **{time_index[0]: index[(index >= start) & (index < end.to_numpy())]})
+                    **{time_index[0]: index[(index >= start) & (index < end)]})
         else:
             self.finished = True
             if buffer_element is not None:
@@ -218,11 +220,13 @@ class BaseStep(ABC):
     def _input_stopped(input_data):
         return (input_data is not None and len(input_data) > 0 and any(map(lambda x: x is None, input_data.values())))
 
-    def reset(self):
+    def reset(self, keep_buffer=False):
         """
         Resets all information of the step concerning a specific run.
+        :param keep_buffer: Flag indicating if the buffer should be resetted too.
         """
-        self.buffer = {}
+        if not keep_buffer:
+            self.buffer = {}
         self.finished = False
         self.current_run_setting = self.default_run_setting.clone()