+++ pywatts/core/run_setting.py
@@ -14,9 +14,12 @@ class RunSetting:
     :type summary_formatter: SummaryFormatter
     """
 
-    def __init__(self, computation_mode: ComputationMode, summary_formatter: SummaryFormatter = SummaryMarkdown()):
+    def __init__(self, computation_mode: ComputationMode, summary_formatter: SummaryFormatter = SummaryMarkdown(),
+                 online_start=None, return_summary=False):
         self.computation_mode = computation_mode
         self.summary_formatter = summary_formatter
+        self.online_start = online_start
+        self.return_summary = return_summary
 
     def update(self, run_setting: 'RunSetting') -> 'RunSetting':
         """
@@ -31,6 +34,8 @@ class RunSetting:
         if setting.computation_mode == ComputationMode.Default:
             setting.computation_mode = run_setting.computation_mode
         setting.summary_formatter = run_setting.summary_formatter
+        setting.online_start = run_setting.online_start
+        setting.return_summary = run_setting.return_summary
         return setting
 
     def clone(self) -> 'RunSetting':
@@ -41,7 +46,9 @@ class RunSetting:
         """
         return RunSetting(
             computation_mode=self.computation_mode,
-            summary_formatter=self.summary_formatter
+            summary_formatter=self.summary_formatter,
+            online_start=self.online_start,
+            return_summary=self.return_summary
         )
 
     def save(self) -> Dict: