+++ pywatts/core/step.py
@@ -1,5 +1,6 @@
 import logging
 import time
+import warnings
 from typing import Optional, Dict, Union, Callable, List
 
 import cloudpickle
@@ -8,6 +9,7 @@ import xarray as xr
 
 from pywatts.callbacks import BaseCallback
 from pywatts.core.base import Base, BaseEstimator
+from pywatts.core.base_condition import BaseCondition
 from pywatts.core.base_step import BaseStep
 from pywatts.core.run_setting import RunSetting
 from pywatts.core.computation_mode import ComputationMode
@@ -38,8 +40,9 @@ class Step(BaseStep):
     :type callbacks: List[Union[BaseCallback, Callable[[Dict[str, xr.DataArray]], None]]]
     :param condition: A callable which checks if the step should be executed with the current data.
     :type condition: Callable[xr.DataArray, xr.DataArray, bool]
-    :param train_if: A callable which checks if the train_if step should be executed or not.
-    :type train_if: Callable[xr.DataArray, xr.DataArray, bool]
+    :param refit_conditions: A List of Callables of BaseConditions, which contains a condition that indicates if
+                                 the module should be trained or not
+    :type refit_conditions: List[Union[BaseCondition, Callable]]
     :param lag: Needed for online learning. Determines what data can be used for retraining.
             E.g., when 24 hour forecasts are performed, a lag of 24 hours is needed, else the retraining would
             use future values as target values.
@@ -54,7 +57,7 @@ class Step(BaseStep):
                  callbacks: List[Union[BaseCallback, Callable[[Dict[str, xr.DataArray]], None]]] = [],
                  condition=None,
                  batch_size: Optional[None] = None,
-                 train_if=None,
+                 refit_conditions=[],
                  retrain_batch=pd.Timedelta(hours=24),
                  lag=pd.Timedelta(hours=24)):
         super().__init__(input_steps, targets, condition=condition,
@@ -64,8 +67,13 @@ class Step(BaseStep):
         self.retrain_batch = retrain_batch
         self.callbacks = callbacks
         self.batch_size = batch_size
+        if self.current_run_setting.computation_mode is not ComputationMode.Refit and len(refit_conditions) > 0:
+            message = "You added a refit_condition without setting the computation_mode to refit." \
+                      " The condition will be ignored."
+            warnings.warn(message)
+            logger.warning(message)
         self.lag = lag
-        self.train_if = train_if
+        self.refit_conditions = refit_conditions
         self.result_steps: Dict[str, ResultStep] = {}
 
     def _fit(self, inputs: Dict[str, BaseStep], target_step):
@@ -75,13 +83,20 @@ class Step(BaseStep):
     def _callbacks(self):
         # plots and writs the data if the step is finished.
         for callback in self.callbacks:
+            dim = _get_time_indexes(self.buffer)[0]
+
+            if self.current_run_setting.online_start is not None:
+                to_plot = {k: self.buffer[k][self.buffer[k][dim] >= self.current_run_setting.online_start] for k in
+                           self.buffer.keys()}
+            else:
+                to_plot = self.buffer
             if isinstance(callback, BaseCallback):
                 callback.set_filemanager(self.file_manager)
             if isinstance(self.buffer, xr.DataArray) or isinstance(self.buffer, xr.Dataset):
                 # DEPRECATED: direct DataArray or Dataset passing is depricated
                 callback({"deprecated": self.buffer})
             else:
-                callback(self.buffer)
+                callback(to_plot)
 
     def _transform(self, input_step):
         if isinstance(self.module, BaseEstimator) and not self.module.is_fitted:
@@ -107,11 +122,11 @@ class Step(BaseStep):
                 condition = cloudpickle.load(pickle_file)
         else:
             condition = None
-        if stored_step["train_if"]:
-            with open(stored_step["train_if"], 'rb') as pickle_file:
-                train_if = cloudpickle.load(pickle_file)
-        else:
-            train_if = None
+        refit_conditions = []
+        for refit_condition in stored_step["refit_conditions"]:
+            with open(refit_condition, 'rb') as pickle_file:
+                refit_conditions.append(cloudpickle.load(pickle_file))
+
         callbacks = []
         for callback_path in stored_step["callbacks"]:
             with open(callback_path, 'rb') as pickle_file:
@@ -120,7 +135,7 @@ class Step(BaseStep):
             callbacks.append(callback)
 
         step = cls(module, inputs, targets=targets, file_manager=file_manager, condition=condition,
-                   train_if=train_if, callbacks=callbacks, batch_size=stored_step["batch_size"])
+                   refit_conditions=refit_conditions, callbacks=callbacks, batch_size=stored_step["batch_size"])
         step.default_run_setting = RunSetting.load(stored_step["default_run_setting"])
         step.current_run_setting = step.default_run_setting.clone()
         step.id = stored_step["id"]
@@ -171,16 +186,17 @@ class Step(BaseStep):
     def get_json(self, fm: FileManager):
         json = super().get_json(fm)
         condition_path = None
-        train_if_path = None
+        refit_conditions_paths = []
         callbacks_paths = []
         if self.condition:
             condition_path = fm.get_path(f"{self.name}_condition.pickle")
             with open(condition_path, 'wb') as outfile:
                 cloudpickle.dump(self.condition, outfile)
-        if self.train_if:
-            train_if_path = fm.get_path(f"{self.name}_train_if.pickle")
-            with open(train_if_path, 'wb') as outfile:
-                cloudpickle.dump(self.train_if, outfile)
+        for refit_condition in self.refit_conditions:
+            refit_conditions_path = fm.get_path(f"{self.name}_refit_conditions.pickle")
+            with open(refit_conditions_path, 'wb') as outfile:
+                cloudpickle.dump(refit_condition, outfile)
+            refit_conditions_paths.append(refit_conditions_path)
         for callback in self.callbacks:
             callback_path = fm.get_path(f"{self.name}_callback.pickle")
             with open(callback_path, 'wb') as outfile:
@@ -188,7 +204,7 @@ class Step(BaseStep):
             callbacks_paths.append(callback_path)
         json.update({"callbacks": callbacks_paths,
                      "condition": condition_path,
-                     "train_if": train_if_path,
+                     "refit_conditions": refit_conditions_paths,
                      "batch_size": self.batch_size})
         return json
 
@@ -200,13 +216,24 @@ class Step(BaseStep):
         """
         if self.current_run_setting.computation_mode in [ComputationMode.Refit] and isinstance(self.module,
                                                                                                BaseEstimator):
-            if self.train_if:
-                input_data = self._get_input(start, end)
-                target = self._get_target(start, end)
-                if self.train_if(input_data, target):
-                    refit_input = self._get_input(end - self.retrain_batch, end)
-                    refit_target = self._get_target(end - self.retrain_batch, end)
-                    self.module.refit(**refit_input, **refit_target)
+            for refit_condition in self.refit_conditions:
+                if isinstance(refit_condition, BaseCondition):
+                    condition_input = {key: value.step.get_result(start, end) for key, value in
+                                       refit_condition.kwargs.items()}
+                    if refit_condition.evaluate(**condition_input):
+                        self._refit(end)
+                        break
+                elif isinstance(refit_condition, Callable):
+                    input_data = self._get_input(start, end)
+                    target = self._get_target(start, end)
+                    if refit_condition(input_data, target):
+                        self._refit(end)
+                        break
+
+    def _refit(self, end):
+        refit_input = self._get_input(end - self.retrain_batch, end)
+        refit_target = self._get_target(end - self.retrain_batch, end)
+        self.module.refit(**refit_input, **refit_target)
 
     def get_result_step(self, item: str):
         if item not in self.result_steps: