+++ tests/unit/core/test_pipeline.py
@@ -21,18 +21,14 @@ pipeline_json = {'id': 1,
                               'is_fitted': False,
                               'module': 'pywatts.modules.wrappers.sklearn_wrapper',
                               'name': 'StandardScaler',
-                              'params': {'copy': True, 'with_mean': True, 'with_std': True},
-                              'sklearn_module': os.path.join('test_pipeline', 'StandardScaler.pickle')},
+                              'sklearn_module': os.path.join('test_pipeline', 'StandardScaler.pickle'),
+                              'targets': []},
                              {'class': 'SKLearnWrapper',
                               'is_fitted': False,
                               'module': 'pywatts.modules.wrappers.sklearn_wrapper',
                               'name': 'LinearRegression',
-                              'params': {'copy_X': True,
-                                         'fit_intercept': True,
-                                         'n_jobs': None,
-                                         'positive': False,
-                                         'normalize': 'deprecated'},
-                              'sklearn_module': os.path.join('test_pipeline', 'LinearRegression.pickle')}],
+                              'sklearn_module': os.path.join('test_pipeline', 'LinearRegression.pickle'),
+                              'targets': []}],
                  'steps': [{'class': 'StartStep',
                             'default_run_setting': {'computation_mode': 4},
                             'id': 1,
@@ -54,7 +50,7 @@ pipeline_json = {'id': 1,
                             'module_id': 0,
                             'name': 'StandardScaler',
                             'target_ids': {},
-                            'train_if': None},
+                            'refit_conditions': []},
                            {'batch_size': None,
                             'callbacks': [],
                             'class': 'Step',
@@ -67,7 +63,7 @@ pipeline_json = {'id': 1,
                             'module_id': 1,
                             'name': 'LinearRegression',
                             'target_ids': {},
-                            'train_if': None}],
+                            'refit_conditions': []}],
                  'version': 1}
 
 
@@ -147,7 +143,7 @@ class TestPipeline(unittest.TestCase):
         assert kwargs["obj"]["steps"] == pipeline_json["steps"]
 
     @patch('pywatts.core.pipeline.FileManager')
-    @patch('pywatts.modules.sklearn_wrapper.pickle')
+    @patch('pywatts.modules.sklearn_wrapper.cloudpickle')
     @patch('pywatts.core.pipeline.json')
     @patch("builtins.open", new_callable=mock_open)
     @patch('pywatts.core.pipeline.os.path.isdir')
@@ -223,8 +219,8 @@ class TestPipeline(unittest.TestCase):
 
         for step in self.pipeline.id_to_step.values():
             assert step.current_run_setting.computation_mode == ComputationMode.FitTransform
-
-        create_summary_mock.assert_has_calls([call(summary_formatter_mock), call(summary_formatter_mock)])
+        assert 2 == create_summary_mock.call_count
+        create_summary_mock.assert_has_calls([call(summary_formatter_mock, None), call(summary_formatter_mock, None)], any_order=True)
 
     @patch('pywatts.core.pipeline.FileManager')
     def test_add_pipeline_to_pipeline_and_test(self, fm_mock):
@@ -386,6 +382,7 @@ class TestPipeline(unittest.TestCase):
         pipeline.start_steps["foo"][0].last = False
         step_one.further_elements.side_effect = [True, True, True, True, False]
         pipeline.add(module=step_one, input_ids=[1])
+        pipeline.current_run_setting = RunSetting(computation_mode=ComputationMode.Transform)
 
         result = pipeline.transform(foo=da)
 
@@ -404,6 +401,7 @@ class TestPipeline(unittest.TestCase):
         step_two.name = "mock"
         step_two.get_result.return_value = {"mock": result_mock}
         self.pipeline.add(module=step_two, input_ids=[1])
+        self.pipeline.current_run_setting = RunSetting(computation_mode=ComputationMode.Transform)
 
         result = self.pipeline.transform(x=input_mock)
 
@@ -459,6 +457,7 @@ class TestPipeline(unittest.TestCase):
         pipeline.start_steps["foo"] = StartStep("foo"), None
         pipeline.start_steps["foo"][0].last = False
         step_one.further_elements.side_effect = [True, True, True, True, True, True, True, False]
+        pipeline.current_run_setting = RunSetting(computation_mode=ComputationMode.Transform)
         pipeline.add(module=step_one, input_ids=[1])
 
         result = pipeline.transform(foo=da)
@@ -581,3 +580,26 @@ class TestPipeline(unittest.TestCase):
         self.pipeline.refit(pd.Timestamp("2000.01.02"), pd.Timestamp("2022.01.02"))
 
         first_step.refit.assert_called_once_with(pd.Timestamp("2000.01.01"), pd.Timestamp("2022.01.01"))
+
+    def test_online_with_filled_buffer(self):
+
+        time = pd.date_range('2000-01-01', freq='1D', periods=7)
+        da = xr.DataArray([2, 3, 4, 3, 3, 1, 2], dims=["time"], coords={'time': time})
+        pipeline = Pipeline(batch=pd.Timedelta(days=1))
+        step_one = MagicMock()
+        step_one.get_result.return_value = {"step": da}
+
+        pipeline.start_steps["foo"] = StartStep("foo"), None
+        pipeline.start_steps["foo"][0].last = False
+        step_one.further_elements.side_effect = [True, True, True, True, False]
+        pipeline.add(module=step_one, input_ids=[1])
+
+        pipeline.test(pd.DataFrame({"foo": [1, 2, 2, 3, 4,5,6]},
+                      index=pd.DatetimeIndex(pd.date_range('2000-01-01', freq='24H', periods=7))),
+                      online_start=pd.to_datetime('2000-01-04'))
+
+        self.assertEqual(5, step_one.get_result.call_count)
+        self.assertEqual(2, step_one.reset.call_count)
+        reset_calls = [call(), call(keep_buffer=True)]
+        step_one.reset.assert_has_calls(reset_calls)
+        self.assertEqual(5, step_one.further_elements.call_count)